move InfoTooltipContent out
